Clean utils.dataset
