+++ definitions.py
@@ -1,0 +1,3 @@
+"""
+Define all important constants in this file.
+"""
