+++ prepare_data.py
@@ -1,0 +1,3 @@
+"""
+Take data from all the CSV files in the input folder and
+"""
