+++ ranking_classes.py
@@ -6,9 +6,9 @@ import os
 from pathlib import Path
 import pandas as pd
 import numpy as np
-import helpfunctions.helpfunctions_dataset as hf_d
-import helpfunctions.helpfunctions_plotly as hf_p
-import helpfunctions.helpfunctions_excel as hf_e
+import utils.dataset as hf_d
+import utils.plotting as hf_p
+import utils.excel as hf_e
 import algos.rank_diff_functions as rdf
 import algos.game_weight_functions as gwf
 import algos.rank_fit_functions as rff
